bbpp/requirements-dev.txt
pytest
pytest-mock
pytest-httpx
pytest-cov
pytest-asyncio